backend/tests/requirements_test.txt
# Dependências para testes
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0

